requirements.txt
asgiref==3.8.1
Django==5.1.2
django-safedelete==1.4.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
